+++ src/app/modules/lux-layout/lux-app-header/lux-app-header-subcomponents/lux-app-header-action-nav/lux-app-header-action-nav-item/lux-app-header-action-nav-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, ContentChild, EventEmitter, Input, Output } from '@angular/core';
-import { LuxActionColorType } from "../../../../../lux-action/lux-action-model/lux-action-component-base.class";
+import { LuxThemePalette } from '../../../../../lux-util/lux-colors.enum';
 import { LuxAppHeaderActionNavItemCustomComponent } from "./lux-app-header-action-nav-item-custom.component";
 
 @Component({
@@ -10,7 +10,7 @@ import { LuxAppHeaderActionNavItemCustomComponent } from "./lux-app-header-actio
 export class LuxAppHeaderActionNavItemComponent {
   @Input() luxLabel = '';
   @Input() luxIconName?: string;
-  @Input() luxColor: LuxActionColorType;
+  @Input() luxColor: LuxThemePalette;
   @Input() luxDisabled = false;
   @Input() luxTagId?: string;