+++ src/app/modules/lux-layout/lux-stepper/lux-stepper-model/lux-stepper-button-config.interface.ts
@@ -1,11 +1,11 @@
-import { LuxActionColorType } from '../../../lux-action/lux-action-model/lux-action-component-base.class';
-
 /**
  * Kennzeichnet die möglichen Optionen, um die Navigations-Buttons des Steppers zu konfigurieren.
  */
+import { LuxThemePalette } from '../../../lux-util/lux-colors.enum';
+
 export interface ILuxStepperButtonConfig {
   label?: string;
-  color?: LuxActionColorType;
+  color?: LuxThemePalette;
   iconName?: string;
   alignIconWithLabel?: boolean;
 }